+++ inst/rmarkdown/templates/rnaseq/skeleton/DE/load_data.R
@@ -1,63 +1,100 @@
 library(tidyverse)
 library(SummarizedExperiment)
 library(janitor)
-load_metrics <- function(se_object, multiqc_data_dir){
+load_metrics <- function(se_object, multiqc_data_dir, gtf_fn, counts){
 
   # bcbio input
   if (!is.na(se_object)){
 
     se <- readRDS(se_object)
-    metrics <- metadata(se)$metrics %>% mutate(sample = toupper(sample)) %>% as.data.frame()
+    metrics <- metadata(se)$metrics %>% as.data.frame()
     # left_join(coldata %>% rownames_to_column('sample')) %>% column_to_rownames('sample')
-
   } else { #nf-core input
 
-    metrics <- read_tsv(paste0(multiqc_data_dir, 'multiqc_general_stats.txt'))
-    metrics_qualimap <- read_tsv(paste0(multiqc_data_dir, 'mqc_qualimap_genomic_origin_1.txt'))
+    # Get metrics from nf-core into bcbio like table
+    # many metrics are already in the Genereal Table of MultiQC, this reads the file
+    metrics <- read_tsv(file.path(multiqc_data_dir, 'multiqc_general_stats.txt'))
 
+    # we get some more metrics from Qualimap and rename columns
+    metrics_qualimap <- read_tsv(file.path(multiqc_data_dir, 'mqc_qualimap_genomic_origin_1.txt'))
     metrics <- metrics %>% full_join(metrics_qualimap)
     metrics <- metrics %>%
       clean_names() %>%
       dplyr::rename_with(~gsub('.*mqc_generalstats_', '', .))
 
+    # This uses the fastqc metrics to get total reads
     total_reads <- metrics %>%
-      filter(grepl('_', sample)) %>%
+      dplyr::filter(!is.na(fastqc_raw_total_sequences)) %>%
       remove_empty(which = 'cols') %>%
       dplyr::rename(single_sample = sample) %>%
       mutate(sample = gsub('_[12]+$', '', single_sample)) %>%
       group_by(sample) %>%
       summarize(total_reads = sum(fastqc_raw_total_sequences))
 
-    if (!("custom_content_biotype_counts_percent_r_rna" %in% colnames(metrics))){
-      metrics[["custom_content_biotype_counts_percent_r_rna"]] <- rep(0,nrow(metrics))
-    }
-    # browser()
-
+    # This renames to user-friendly names the metrics columns
     metrics <- metrics %>%
-      filter(!grepl('_[12]$', sample)) %>%
+      dplyr::filter(is.na(fastqc_raw_total_sequences)) %>%
       remove_empty(which = 'cols') %>%
       full_join(total_reads) %>%
-      dplyr::rename(mapped_reads = samtools_reads_mapped) %>%
+      mutate(mapped_reads = samtools_reads_mapped) %>%
       mutate(exonic_rate = exonic/(star_uniquely_mapped * 2)) %>%
       mutate(intronic_rate = intronic/(star_uniquely_mapped * 2)) %>%
       mutate(intergenic_rate = intergenic/(star_uniquely_mapped * 2)) %>%
-      dplyr::rename(r_rna_rate = custom_content_biotype_counts_percent_r_rna) %>%
-      mutate(x5_3_bias = qualimap_5_3_bias) %>%
-      dplyr::select(
-        sample,
-        total_reads,
-        mapped_reads,
-        exonic_rate,
-        intronic_rate,
-        intergenic_rate,
-        r_rna_rate,
-        x5_3_bias
-      ) %>%
-      dplyr::select(where(~!all(is.na(.)))) %>%
-      as.data.frame()
+      mutate(x5_3_bias = qualimap_5_3_bias)
 
-    # metrics <- metrics %>%
-    #   full_join(meta_df , by = c("sample" = "sample"))
+    # Sometimes we don't have rRNA due to mismatch annotation, We skip this if is the case
+    gtf <- NULL
+    if (genome =="other"){
+      gtf <- gtf_fn
+    }else{
+      if (genome == "hg38") {
+        gtf <- "hg38.rna.gtf.gz"
+      } else if (genome == "mm10") {
+        gtf <- "mm10.rna.gtf.gz"
+      } else if (genome == "mm39") {
+        gtf <- "mm39.rna.gtf.gz"
+      }
+      gtf <- system.file("extdata", "annotation",
+                         gtf,
+                         package="bcbioR")
+    }
+    if (is.null(gtf)) {
+      print("No genome provided! Please add it at the top of this Rmd")
+    }
+
+    gtf=rtracklayer::import(gtf)
+
+
+    one=grep("gene_type", colnames(as.data.frame(gtf)), value = TRUE)
+    another=grep("gene_biotype", colnames(as.data.frame(gtf)), value = TRUE)
+    biotype=NULL
+    if(length(one)==1){
+      biotype=one
+    }else if(length(another)==1){
+      biotype=another
+    }else{
+      warning("No gene biotype founded")
+    }
+
+    if (!is.null(biotype)){
+      annotation=as.data.frame(gtf) %>% .[,c("gene_id", biotype)]
+      rRNA=grepl("rRNA|tRNA",annotation[[biotype]])
+      genes=intersect(annotation[rRNA,"gene_id"],row.names(counts))
+      ratio=data.frame(sample=colnames(counts),
+                       r_and_t_rna_rate=colSums(counts[genes,])/colSums(counts))
+      metrics = left_join(metrics, ratio, by="sample")
+    }else{
+      metrics[["r_and_t_rna_rate"]] <- NA
+    }
+
+    # if ("custom_content_biotype_counts_percent_r_rna" %in% colnames(metrics)){
+    #   metrics <- mutate(metrics, r_rna_rate = custom_content_biotype_counts_percent_r_rna)
+    # }else{
+    #  metrics[["r_rna_rate"]] <- NA
+    # }
+    metrics=metrics[,c("sample","mapped_reads","exonic_rate","intronic_rate",
+                       "total_reads",
+                       "x5_3_bias", "r_and_t_rna_rate","intergenic_rate")]
   }
   metrics$sample <- make.names(metrics$sample)
   rownames(metrics) <- metrics$sample
@@ -69,7 +106,7 @@ load_coldata <- function(coldata_fn, column, numerator, denominator, subset_colu
     dplyr::select(!matches("fastq") & !matches("strandness")) %>%
     distinct()
   if('description' %in% names(coldata)){
-    coldata$sample <- coldata$description
+    coldata$sample <- tolower(coldata$description)
   }
   coldata <- coldata %>% distinct(sample, .keep_all = T)
   stopifnot(column %in% names(coldata))
@@ -93,10 +130,14 @@ load_counts <- function(counts_fn){
 
   # bcbio input
   if(grepl('csv', counts_fn)){
-    counts <- read_csv(counts_fn) %>% column_to_rownames('gene')
+    counts <- read_csv(counts_fn) %>%
+      mutate(gene = str_replace(gene, pattern = "\\.[0-9]+$", "")) %>%
+      column_to_rownames('gene')
+    colnames(counts) = tolower(colnames(counts))
+    return(counts)
   } else { # nf-core input
     counts <- read_tsv(counts_fn) %>% dplyr::select(-gene_name) %>%
-      mutate(gene_id = str_replace(gene_id, pattern = "\\.[0-9]$", "")) %>%
+      mutate(gene_id = str_replace(gene_id, pattern = "\\.[0-9]+$", "")) %>%
       column_to_rownames('gene_id') %>% round
 
     return(counts)