+++ inst/rmarkdown/templates/teaseq/skeleton/QC/QC-01-load_data.R
@@ -1,3 +1,12 @@
+library(Seurat)
+library(tidyverse)
+library(Matrix)
+library(data.table)
+library(magrittr)
+library(Signac)
+library(EnsDb.Hsapiens.v86)
+library(qs)
+library(bcbioR)
 # prepare seurat object with all 3 modes
 
 # read cellranger-arc outs